Move load to S3 functions into load.py.
