+++ report/report.py
@@ -1,7 +1,5 @@
 """Script for creating PDF summary report."""
 
-from os import environ as ENV
-
 from datetime import datetime as dt
 from email.mime.multipart import MIMEMultipart
 from email.mime.application import MIMEApplication
@@ -13,8 +11,6 @@ from reportlab.lib import colors
 from reportlab.lib.styles import getSampleStyleSheet
 from pandas import DataFrame
 from dotenv import load_dotenv
-from boto3 import client
-from botocore.exceptions import ClientError
 
 from extract_report_data import get_days_data_per_station, get_db_connection
 from transform_summary import get_station_summary
@@ -59,21 +55,6 @@ def get_email_message_as_string(station_name: str, pdf_bytes: bytes) -> str:
     msg.attach(attachment)
 
     return msg
-
-
-def get_s3_client() -> client:
-    """Returns an S3 client using boto3."""
-
-    s3_client = client(
-        "s3", aws_access_key_id=ENV["AWS_ACCESS_KEY_ID"],
-        aws_secret_access_key=ENV["AWS_SECRET_ACCESS_KEY"])
-
-    return s3_client
-
-
-def report_already_exists(bucket_name: str, filename: str) -> bool:
-    """Returns true if a file with the given filename already exists in S3 bucket."""
-    ...
 
 
 if __name__ == "__main__":
